# src/main/java/org/bitebuilders/controller/dto/MessageResponseDTO.java
package org.bitebuilders.controller.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.ToString;

@Data
@AllArgsConstructor
@ToString
public class MessageResponseDTO {
    private String message;
}