# src/main/java/org/bitebuilders/controller/requests/StartEventRequest.java
package org.bitebuilders.controller.requests;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Data
@Getter
@NoArgsConstructor
@AllArgsConstructor
public class StartEventRequest {
    private Long eventId;
}