# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CRM module for Uralintern",
    "description": "API documentation for Uralintern CRM Module.\n\n**Контроллеры:**\n- **auth-controller**: Отвечает за регистрацию и авторизацию пользователей, включая регистрацию с реферальной ссылкой и обновление токенов.\n- **user-info-controller**: Управляет пользовательскими данными, включая обновление профиля, получение реферальных ссылок и список руководителей.\n- **event-controller**: Обеспечивает управление мероприятиями, включая их создание, обновление, скрытие, удаление, а также доступ к информации о мероприятиях.\n- **message-controller**: Управляет шаблонами сообщений для мероприятий, включая их получение и обновление.\n- **event-student-controller**: Обеспечивает управление заявками и участием студентов в мероприятиях, включая отправку заявок, их принятие/отклонение и управление студентами на мероприятии.\n- **event-curator-controller**: Управляет заявками и участием кураторов в мероприятиях, включая их принятие/отклонение и управление кураторами.\n- **notification-controller**: Управляет уведомлениями пользователей, связанными с их участием в мероприятиях.\n",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost/api",
      "description": "kayakto server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/events/post": {
      "post": {
        "tags": [
          "event-controller"
        ],
        "summary": "пока не работает",
        "operationId": "createEvent",
        "description": "Создание нового мероприятия. Доступно только администраторам.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Мероприятие успешно создано.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/start-event": {
      "post": {
        "tags": [
          "event-controller"
        ],
        "summary": "пока не работает",
        "operationId": "startEvent",
        "description": "Досрочно начать мероприятие. Доступ - администратор и руководитель проекта.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/{eventId}": {
      "get": {
        "tags": [
          "event-controller"
        ],
        "operationId": "getEventByID",
        "description": "Получение данных мероприятия по его идентификатору. Доступ - администратор или руководитель.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/my": {
      "get": {
        "tags": [
          "event-controller"
        ],
        "operationId": "getMyEvents",
        "description": "Получение данных мероприятий, с которыми связан пользователь, отправивший запрос. Доступ: все авторизованные пользователи.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDTO"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/active": {
      "get": {
        "tags": [
          "event-controller"
        ],
        "operationId": "getOpenedEvents",
        "description": "Получение мероприятий, доступных для регистрации. Доступно всем авторизованным пользователям.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDTO"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/update/{eventId}": {
      "put": {
        "tags": [
          "event-controller"
        ],
        "summary": "пока не работает",
        "operationId": "updateEvent",
        "description": "Обновление информации о мероприятии. Доступно администраторам связанным с мероприятием.\n",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "ID мероприятия для обновления.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Мероприятие успешно обновлено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/hide/{eventId}": {
      "put": {
        "summary": "пока не работает",
        "tags": [
          "event-controller"
        ],
        "operationId": "hideEvent",
        "description": "Поменять статус мероприятия на \"скрытое\". Или вернуть состояние мероприятия к обычному. Доступ - руководитель и администратор.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events/delete/{eventId}": {
      "delete": {
        "tags": [
          "event-controller"
        ],
        "operationId": "deleteEvent",
        "description": "Удаление мероприятия. Доступно только администраторам.\n",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "ID мероприятия для удаления.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Мероприятие успешно удалено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "BadRequestResponse": {
        "description": "Bad Request, запрос неверный со стороны клиента",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid argument"
                }
              }
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "Forbidden, отказано в доступе",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Access denied"
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "Not Found, такого элемента не существует",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Event with id 123 not found"
                }
              }
            }
          }
        }
      },
      "InternalServerErrorResponse": {
        "description": "Internal Server Error, запрос вызвал ошибку на сервере",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "EventRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "CRM модуль"
          },
          "descriptionText": {
            "type": "string",
            "example": "CRM модуль для мастеров глиняной лепки"
          },
          "adminId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "managerId": {
            "type": "integer",
            "format": "int64",
            "example": 2
          },
          "eventStartDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-22T22:00:00+05:00"
          },
          "eventEndDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-18T22:00:00+05:00"
          },
          "enrollmentStartDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T22:00:00+05:00"
          },
          "enrollmentEndDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-22T22:00:00+05:00"
          },
          "chatUrl": {
            "type": "string",
            "example": "https://t.me/chaturl"
          },
          "numberSeatsStudent": {
            "type": "integer",
            "format": "int32",
            "example": 50
          },
          "hasTest": {
            "type": "boolean",
            "example": false
          },
          "testUrl": {
            "type": "string",
            "example": null
          }
        }
      },
...
      "EventDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "condition": {
            "type": "string",
            "enum": [
              "PREPARATION",
              "REGISTRATION_OPEN",
              "NO_SEATS",
              "REGISTRATION_CLOSED",
              "IN_PROGRESS",
              "FINISHED",
              "HIDDEN",
              "DELETED"
            ],
            "example": "REGISTRATION_OPEN"
          },
          "descriptionText": {
            "type": "string",
            "example": "CRM модуль для мастеров глиняной лепки"
          },
          "title": {
            "type": "string",
            "example": "CRM модуль"
          },
          "adminId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "managerId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "eventStartDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-22T22:00:00+05:00"
          },
          "eventEndDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-18T22:00:00+05:00"
          },
          "enrollmentStartDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T22:00:00+05:00"
          },
          "enrollmentEndDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-22T22:00:00+05:00"
          },
          "numberSeatsStudent": {
            "type": "integer",
            "format": "int32",
            "example": 50
          },
          "chatUrl": {
            "type": "string",
            "example": "https://t.me/chaturl"
          },
          "hasTest": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "StartEventRequest": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "integer",
            "format": "int64",
            "example": 123
          }
        }
      },
...
      "MessageResponseDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}